dashboard/requirements.txt
streamlit
google-cloud-bigquery
pandas
plotly
db-dtypes
